feat!: use shiki to highlight code and move to ESM
