+++ lib/locator.js
@@ -1,8 +1,7 @@
 /* eslint-disable no-magic-numbers */
-const chalk = require( 'chalk' )
-const emphasize = require( 'emphasize' )
+import chalk from 'chalk'
 
-function locate( source, locations ) {
+function locate( source, locations, highlight ) {
   let lines = source.split( '\n' )
 
   const frameStart = 0
@@ -15,16 +14,7 @@ function locate( source, locations ) {
 
   const digits = String( frameEnd ).length + 1
 
-  const formattedLines = emphasize.highlight( 'js', source, {
-    keyword: chalk.magenta,
-    built_in: chalk.cyan, // eslint-disable-line
-    string: chalk.green,
-    number: chalk.green,
-    literal: chalk.green,
-    'class': chalk.yellow, // eslint-disable-line
-    'function': chalk.reset.white, // eslint-disable-line
-    params: chalk.reset.white,
-  } ).value.split( '\n' )
+  const formattedLines = highlight(source).split('\n')
 
   return lines
     .map( ( str, i ) => {
@@ -158,4 +148,6 @@ function getErrorLocations( line, locations ) {
   return locations.filter( l => l.line === line )
 }
 
-exports.locate = locate
+export {
+  locate
+}
