+++ package.json
@@ -1,17 +1,15 @@
 {
   "name": "eslint-formatter-mo",
   "description": "Good-lookin' ESLint formatter for delightful readability",
-  "version": "1.3.1",
-  "main": "lib/index.js",
+  "version": "2.0.0-alpha.0",
+  "type": "module",
+  "exports": "./lib/index.js",
   "files": [
     "lib"
   ],
   "scripts": {
-    "fixture": "eslint fixtures -f ./",
-    "lint": "eslint lib -f ./"
-  },
-  "eslintConfig": {
-    "extends": "mo"
+    "fixture": "eslint fixtures -f mo",
+    "lint": "eslint lib -f mo"
   },
   "repository": {
     "type": "git",
@@ -33,19 +31,29 @@
   },
   "homepage": "https://github.com/fengzilong/eslint-formatter-mo#readme",
   "dependencies": {
-    "chalk": "^3.0.0",
-    "emphasize": "^4.2.0",
-    "gradient-string": "^1.2.0",
-    "plur": "^4.0.0",
-    "string-width": "^4.2.0",
-    "terminal-link": "^2.1.1"
+    "chalk": "^5.4.1",
+    "gradient-string": "^3.0.0",
+    "plur": "^5.1.0",
+    "string-width": "^7.2.0",
+    "terminal-link": "^3.0.0",
+    "shiki": "^1.26.1",
+    "@shikijs/themes": "^1.26.1",
+    "@shikijs/langs": "^1.26.1"
   },
   "devDependencies": {
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
-    "eslint": "^8.1.0",
+    "eslint": "^9",
     "eslint-config-mo": "^0.5.2",
-    "semantic-release": "^19.0.3"
+    "semantic-release": "^19.0.3",
+    "@eslint/js": "^9.17.0",
+    "typescript-eslint": "^8.19.0",
+    "eslint-plugin-import-x": "^4.6.1",
+    "globals": "^15.14.0",
+    "eslint-plugin-simple-import-sort": "^12.1.1"
+  },
+  "engines": {
+    "node": ">=18"
   },
   "packageManager": "pnpm@9.15.1"
 }
